lazy init custom tabs, refactor package names
