+++ app/src/main/java/os/dtakac/feritraspored/schedule/di/ScheduleModule.kt
@@ -4,7 +4,7 @@ import org.koin.android.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 import os.dtakac.feritraspored.schedule.repository.ScheduleRepository
 import os.dtakac.feritraspored.schedule.repository.ScheduleRepositoryImpl
-import os.dtakac.feritraspored.schedule.view_model.ScheduleViewModel
+import os.dtakac.feritraspored.schedule.viewmodel.ScheduleViewModel
 
 val scheduleModule = module {
     factory<ScheduleRepository> { ScheduleRepositoryImpl(get()) }
