+++ app/src/main/java/os/dtakac/feritraspored/settings/di/SettingsModule.kt
@@ -2,7 +2,7 @@ package os.dtakac.feritraspored.settings.di
 
 import org.koin.android.viewmodel.dsl.viewModel
 import org.koin.dsl.module
-import os.dtakac.feritraspored.settings.view_model.PreferenceViewModel
+import os.dtakac.feritraspored.settings.viewmodel.PreferenceViewModel
 
 val settingsModule = module {
     viewModel { PreferenceViewModel(get(), get()) }
